+++ public/jsdocMd.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { readFileSync } = require('fs');
+const fs = require('fs');
 const globby = require('globby');
 const codeToJsdocComments = require('../private/codeToJsdocComments');
 const jsdocCommentToMember = require('../private/jsdocCommentToMember');
@@ -17,6 +17,7 @@ const membersToMdAst = require('../private/membersToMdAst');
  * @param {string} [options.sourceGlob='**\/*.{mjs,js}'] JSDoc source file glob pattern.
  * @param {string} [options.markdownPath='readme.md'] Path to the markdown file for docs insertion.
  * @param {string} [options.targetHeading='API'] Markdown file heading to insert docs under.
+ * @returns {Promise<void>} Resolves once the operation is complete.
  * @example <caption>Ways to `import`.</caption>
  * ```js
  * import { jsdocMd } from 'jsdoc-md';
@@ -42,32 +43,50 @@ const membersToMdAst = require('../private/membersToMdAst');
  *   sourceGlob: 'index.mjs',
  *   markdownPath: 'README.md',
  *   targetHeading: 'Docs',
+ * }).then(() => {
+ *   console.log('Done!');
  * });
  * ```
  */
-module.exports = function jsdocMd({
+module.exports = async function jsdocMd({
   cwd = process.cwd(),
   sourceGlob = '**/*.{mjs,js}',
   markdownPath = 'readme.md',
   targetHeading = 'API',
 } = {}) {
-  const members = [];
+  const codeFilePaths = await globby(sourceGlob, { cwd, gitignore: true });
+  const codeFiles = new Map();
+  const jsdocMembers = [];
 
-  for (const codeFilePath of globby.sync(sourceGlob, {
-    cwd,
-    gitignore: true,
-  })) {
-    const code = readFileSync(codeFilePath, { encoding: 'utf8' });
+  await Promise.all(
+    codeFilePaths.map(async (codeFilePath) => {
+      // Update the code files map.
+      codeFiles.set(
+        codeFilePath,
+        await fs.promises.readFile(codeFilePath, { encoding: 'utf8' })
+      );
 
-    for (const jsdocComment of codeToJsdocComments(code, codeFilePath)) {
-      const member = jsdocCommentToMember(jsdocComment, code, codeFilePath);
-      if (member) members.push(member);
-    }
-  }
+      // Get the JSDoc comments from the code.
+      const jsdocComments = codeToJsdocComments(
+        codeFiles.get(codeFilePath),
+        codeFilePath
+      );
+
+      // Get the JSDoc members from the JSDoc comments.
+      for (const jsdocComment of jsdocComments) {
+        const jsdocMember = jsdocCommentToMember(
+          jsdocComment,
+          codeFiles.get(codeFilePath),
+          codeFilePath
+        );
+        if (jsdocMember) jsdocMembers.push(jsdocMember);
+      }
+    })
+  );
 
   mdFileReplaceSection({
     markdownPath,
     targetHeading,
-    replacementAst: membersToMdAst(members),
+    replacementAst: membersToMdAst(jsdocMembers),
   });
 };